Update
Added new feature of being able to purge current secret key and invalidate all current logins. This means that it is much easier to manage.
Webpage also now reacts to the user account if logged it by displaying whether the user is a student or staff in the page title if logged in, and if not logged in, it displays guest